Changed some problems with old password + changed test to compile with former changes

+++ src/policy/PolicyHandler.java
@@ -168,7 +168,7 @@ public class PolicyHandler {
     public static boolean shouldUserChangePassword(Forum forum, User user) {
         try {
             long passwordSetTime = user.getPasswordSetDate().getTime();
-            int passwordMaxTime = forum.getPolicy().getPasswordMaxTime() * 24 * 60 * 60 * 1000;
+            long passwordMaxTime = ((long) forum.getPolicy().getPasswordMaxTime()) * 24 * 60 * 60 * 1000;
             return (new Date()).after(new Date(passwordSetTime + passwordMaxTime));
         } catch (Exception e) {
             return false;
